reward tests refactor
- put the test code in its own file
